+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Enhancements
 
 - Improve diagnostics for span-related failures [813](https://github.com/bugsnag/maze-runner/pull/813)
+- Write received spans to file after each scenario [814](https://github.com/bugsnag/maze-runner/pull/814)
 
 # v10.7.3 - 2025/11/17